+++ CHANGELOG.md
@@ -3,12 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The changelog adheres to format specified in [keep change log](http://keepachangelog.com)
 
-## [Unreleased]
+## [1.1.0] - 2016-07-23
 ### Added
 - Avatar Image for chats
 - Meta Message to self when you join the room 
-- Group users box (dummy content).
-- Group users slide navigation for mobile devices 
+- List Group users.
+- Group users slide navigation for mobile devices.
 
 ### Changed
 - UI modified : Meta messages removed bg color and added well.
@@ -16,6 +16,7 @@ The changelog adheres to format specified in [keep change log](http://keepachang
 - Included new version of cartoon-avatar.
 - chat input box moved outside 
 - removed extra unwanted div tags.
+- Updated Color values.
 
 ### Fixed
 - chat input box for mozilla was displayed to the right.